move annotations to own package
