+++ src/main/java/com/publicuhc/pluginframework/routing/annotation/CommandMethod.java
@@ -1,5 +1,5 @@
 /*
- * OptionsMethod.java
+ * CommandMethod.java
  *
  * Copyright (c) 2014. Graham Howden <graham_howden1 at yahoo.co.uk>.
  *
@@ -19,12 +19,28 @@
  * along with PluginFramework.  If not, see <http ://www.gnu.org/licenses/>.
  */
 
-package com.publicuhc.pluginframework.routing;
+package com.publicuhc.pluginframework.routing.annotation;
+
+import org.bukkit.command.CommandSender;
 
 import java.lang.annotation.*;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 @Documented
-public @interface OptionsMethod
-{}
+public @interface CommandMethod
+{
+    public static final String DEFAULT_HELP = "?";
+
+    public String command();
+
+    public boolean options() default false;
+
+    public String[] optionOrder() default {};
+
+    public String[] permissions() default {};
+
+    public String helpOption() default DEFAULT_HELP;
+
+    public Class<? extends CommandSender>[] allowedSenders() default CommandSender.class;
+}
