+++ src/main/java/com/publicuhc/pluginframework/routing/annotation/OptionsMethod.java
@@ -1,5 +1,5 @@
 /*
- * Permission.java
+ * OptionsMethod.java
  *
  * Copyright (c) 2014. Graham Howden <graham_howden1 at yahoo.co.uk>.
  *
@@ -19,22 +19,12 @@
  * along with PluginFramework.  If not, see <http ://www.gnu.org/licenses/>.
  */
 
-package com.publicuhc.pluginframework.routing;
+package com.publicuhc.pluginframework.routing.annotation;
 
 import java.lang.annotation.*;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.METHOD})
+@Target(ElementType.METHOD)
 @Documented
-public @interface Permission
-{
-    /**
-     * List of permissions to check for the method to run
-     */
-    String[] permissions() default {};
-
-    /**
-     * Whether the method needs all permissions to pass or just one
-     */
-    boolean needsAll() default true;
-}
+public @interface OptionsMethod
+{}
