+++ src/test/java/com/publicuhc/pluginframework/routing/testcommands/SampleCommand.java
@@ -21,8 +21,8 @@
 
 package com.publicuhc.pluginframework.routing.testcommands;
 
-import com.publicuhc.pluginframework.routing.CommandMethod;
-import com.publicuhc.pluginframework.routing.OptionsMethod;
+import com.publicuhc.pluginframework.routing.annotation.CommandMethod;
+import com.publicuhc.pluginframework.routing.annotation.OptionsMethod;
 import joptsimple.OptionDeclarer;
 import joptsimple.OptionSet;
 import org.bukkit.command.CommandSender;
